Break up Editor.ts into focused modules
Extract drag/selection logic into DragManager (1241 lines) and
zoom/pan/loupe logic into ZoomManager (541 lines), reducing Editor.ts
from 6879 to 4266 lines. Remove the localization window which is no
longer needed. Incorporate touch auto-scroll sensitivity changes from
main.
